Simplified test client. Removed un-needed features, files,  and OIDC flow. Test panel is now the single page
