+++ examples/clients/web/react/passkey-client/src/App.css
@@ -34,7 +34,7 @@
   border-bottom: 1px solid #000;
   line-height: 0.1em;
   margin: 10px 0 20px;
-} 
+}
 
 .middle-text span {
   background: #fff;
@@ -45,7 +45,7 @@
   padding-top: 1em;
 }
 
-.footer-item > div {
+.footer-item>div {
   margin-top: 1.5em;
   margin: 2em 0;
 }
@@ -66,8 +66,7 @@
   display: block;
 }
 
-.svg-container
-{
+.svg-container {
   display: flex;
   align-items: center;
   background-color: #e9e9e9;
@@ -99,13 +98,13 @@
 
 
 
-.svg-1
-{
+.svg-1 {
   margin: 0 auto;
   display: block;
 }
 
-html, body {
+html,
+body {
   height: 100%;
 }
 
@@ -122,7 +121,35 @@ body {
   from {
     transform: rotate(0deg);
   }
+
   to {
     transform: rotate(360deg);
   }
+}
+
+.reg-item {
+  text-align: center;
+  border-right: 4px solid #d5d5d5;
+}
+
+.auth-item {
+  text-align: center;
+}
+
+.username-form {
+  margin-top: 2em;
+}
+
+.test-section {
+  margin-top: 2em;
+  padding-bottom: 3em;
+  border-bottom: 5px solid black;
+}
+
+.alert {
+  margin-top: 1em;
+}
+
+.card {
+  margin-bottom: 2em;
 }
