+++ examples/clients/web/react/passkey-client/src/components/header.jsx
@@ -1,6 +1,5 @@
 import Navbar from "react-bootstrap/Navbar";
 import Nav from "react-bootstrap/Nav";
-import { Link } from "react-router-dom";
 import Container from "react-bootstrap/Container";
 
 export default function Header() {
@@ -8,22 +7,6 @@ export default function Header() {
     <Navbar bg="dark" variant="dark">
       <Container>
         <Navbar.Brand>Passkey Starter Kit</Navbar.Brand>
-        <Navbar.Collapse id="basic-navbar-nav">
-          <Nav className="me-auto">
-            <Link className="nav-link" to="/">
-              Home
-            </Link>
-            <Link className="nav-link" to="sign_in">
-              Sign In
-            </Link>
-            <Link className="nav-link" to="sign_up">
-              Sign up
-            </Link>
-            <Link className="nav-link" to="test_panel">
-              Test Panel
-            </Link>
-          </Nav>
-        </Navbar.Collapse>
       </Container>
     </Navbar>
   );
