--- examples/clients/web/react/passkey-client/src/components/loading.jsx
@@ -1,10 +1,0 @@
-import Spinner from "react-bootstrap/Spinner";
-
-export default function Loading({ loadingInfo }) {
-  return (
-    <div style={{ textAlign: "center" }}>
-      <Spinner size="lg" />
-      <h5>{loadingInfo.message}</h5>
-    </div>
-  );
-}
