--- examples/clients/web/react/passkey-client/src/services/Utils.js
@@ -1,16 +1,0 @@
-const Utils = {
-  timeoutUtil,
-  convertDate
-}
-
-async function timeoutUtil(timeout) {
-  await new Promise((resolve) => {
-    setTimeout(() => resolve(), timeout);
-  })
-}
-
-function convertDate(myDate) {
-  return new Date(myDate).toLocaleString();
-}
-
-export default Utils;
